% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_features_parallel.R
\name{construct_features_parallel}
\alias{construct_features_parallel}
\title{construct_features_parallel}
\usage{
construct_features_parallel(
  output_path,
  gen = "Hsapiens",
  gen_ver = "hg19",
  sig = "GATC",
  bin_type = "Bins-uniform",
  binsize = 5000,
  wg_file = NULL,
  chrs = NULL,
  feature_type = "RE-based",
  ncore = NULL
)
}
\arguments{
\item{output_path}{the path to the folder and name prefix you want to place
feature files into. The feature file will have the suffix '_bintolen.txt.gz'.}

\item{gen}{name of the species: e.g., default \code{'Hsapiens'}.}

\item{gen_ver}{genomic assembly version: e.g., default \code{'hg19'}.}

\item{sig}{restriction enzyme cut pattern (or a vector of patterns; e.g.,
'GATC' or c('GATC','GANTC')).}

\item{bin_type}{'Bins-uniform' if uniformly binned by binsize in
bp, or 'Bins-RE-sites' if binned by number of
restriction enzyme fragments.}

\item{binsize}{binsize in bp if bin_type='Bins-uniform' (or number of
RE fragment cut sites if bin_type='Bins-RE-sites'), defaults to 5000.}

\item{wg_file}{path to the bigWig file containing mappability values across
the genome of interest.}

\item{chrs}{select a subset of chromosomes' e.g.,
c('chr21','chr22'). Defaults to all chromosomes (except Y and M)
in the genome specified.}

\item{feature_type}{'RE-based' if features are to be computed based on
restriction enzyme fragments. 'RE-agnostic' ignores restriction enzyme cutsite 
information and computes features gc and map based on binwide averages. bin_type
has to be 'Bins-uniform' if \code{feature_type='RE-agnostic'}.}

\item{ncore}{Number of cores to parallelize. Defaults to 
\code{parallel::detectCores()-1}.}
}
\value{
a features 'bintolen' file that contains GC, mappability and length
features.
}
\description{
This function lists all restriction enzyme cutsites of a given genome and
genome version with genomic features outlined in Carty et al. (2017)
https://www.nature.com/articles/ncomms15454; GC content, mappability,
and effective length
}
\examples{
outdir<-paste0(tempdir(check=TRUE),'/')
construct_features_parallel(output_path=outdir,gen='Hsapiens',
gen_ver='hg19',sig=c('GATC','GANTC'),bin_type='Bins-uniform',binsize=100000,
wg_file=NULL,chrs=c('chr21'),ncore=2)
}
