% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chr_sizes.R
\name{get_chr_sizes}
\alias{get_chr_sizes}
\title{get_chr_sizes}
\usage{
get_chr_sizes(gen = "Hsapiens", gen_ver = "hg19", chrs = NULL)
}
\arguments{
\item{gen}{name of the species: e.g., default \code{'Hsapiens'}}

\item{gen_ver}{genomic assembly version: e.g., default \code{'hg19'}}

\item{chrs}{select a subset of chromosomes' e.g.,
c('chr21','chr22'). Defaults to all chromosomes (except Y and M)
in the genome specified.}
}
\value{
named vector containing names as chromosomes and values as chromosome
sizes.
}
\description{
This function finds all chromosome sizes of a given genome, genome version
and set of chromosomes.
}
\examples{
get_chr_sizes('Hsapiens','hg19',c('chr21','chr22'))
}
