% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chrs.R
\name{get_chrs}
\alias{get_chrs}
\title{get_chrs}
\usage{
get_chrs(gen = "Hsapiens", gen_ver = "hg19")
}
\arguments{
\item{gen}{name of the species: e.g., default \code{'Hsapiens'}}

\item{gen_ver}{genomic assembly version: e.g., default \code{'hg19'}}
}
\value{
string vector of chromosomes.
}
\description{
This function finds all chromosomes of a given genome and genome version
except for Y and M.
}
\examples{
get_chrs('Hsapiens','hg19')
}
