% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi.list.Dthreshold.detect.R
\name{gi.list.Dthreshold.detect}
\alias{gi.list.Dthreshold.detect}
\title{gi.list.Dthreshold.detect}
\usage{
gi.list.Dthreshold.detect(gi_list)
}
\arguments{
\item{gi_list}{A valid gi_list instance. See
\code{?gi.list.validate} for more details about
the attributes of a valid gi_list instance.}
}
\value{
maximum genomic distance in the object
}
\description{
This function finds the maximum genomic distance in a valid gi_list object.
}
\examples{
gi_list<-generate.binned.gi.list(1e6,chrs='chr22')
gi.list.Dthreshold.detect(gi_list)
}
