% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi.list.validate.R
\name{gi.list.validate}
\alias{gi.list.validate}
\title{gi.list.validate}
\usage{
gi.list.validate(gi_list)
}
\arguments{
\item{gi_list}{gi_list object to be verified. In order to pass without 
errors,  a gi_list object (1) has to be a list of
InteractionSet::GInteractions objects, (2) each list element has to be named
as chromosomes and only contain intra-chromosomal interaction information,
(3) \code{mcols(.)} for each list element should at least contain pairwise
genomic distances in a column named 'D'.}
}
\value{
invisible value if the gi_list instance is valid. Otherwise, an error
is raised.
}
\description{
This function validates a gi_list instance.
}
\examples{
gi_list<-generate.binned.gi.list(1e6,chrs='chr22')
gi.list.validate(gi_list)
}
