% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi_list_Dthreshold_detect.R
\name{gi_list_Dthreshold.detect}
\alias{gi_list_Dthreshold.detect}
\title{gi_list_Dthreshold_detect}
\usage{
gi_list_Dthreshold.detect(gi_list)
}
\arguments{
\item{gi_list}{A valid gi_list instance. See
\code{?gi_list_validate} for more details about
the attributes of a valid gi_list instance.}
}
\value{
maximum genomic distance in the object
}
\description{
This function finds the maximum genomic distance in a valid gi_list object.
}
\examples{
gi_list<-generate_binned_gi_list(1e6,chrs='chr22')
gi_list_Dthreshold.detect(gi_list)
}
