% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCDCPlus_parallel.R
\name{HiCDCPlus_parallel}
\alias{HiCDCPlus_parallel}
\title{HiCDCPlus_parallel}
\usage{
HiCDCPlus_parallel(
  gi_list,
  covariates = NULL,
  chrs = NULL,
  distance_type = "spline",
  model_distribution = "nb",
  binned = TRUE,
  df = 6,
  Dmin = 0,
  Dmax = 2e+06,
  ssize = 0.01,
  ncore = NULL
)
}
\arguments{
\item{gi_list}{List of \code{GenomicInteractions} objects where each object
named with chromosomes contains intrachromosomal interaction information
(minimally containing counts and genomic distance in 
\code{mcols(gi_list[[1]])}---see
\code{?gi_list_validate} for a detailed explanation of valid \code{gi_list}
instances).}

\item{covariates}{covariates to be considered in addition to genomic
distance D. Defaults to all covariates besides 
'D','counts','mu','sdev',pvalue','qvalue' 
in \code{mcols(gi)}}

\item{chrs}{select a subset of chromosomes' e.g.,
c('chr21','chr22'). Defaults to all chromosomes
in the \code{gi_list}.}

\item{distance_type}{distance covariate form: 'spline' or 'log'.
Defaults to 'spline'.}

\item{model_distribution}{'nb' uses a Negative Binomial model, 
'nb_vardisp' uses a Negative Binomial model with a distance specific 
dispersion parameter inferred from the data, 'nb_hurdle' uses the legacy
HiC-DC model.}

\item{binned}{TRUE if uniformly binned or FALSE if binned by
restriction enzyme fragment cutsites}

\item{df}{degrees of freedom for the genomic distance spline
function if \code{distance_type='spline'}. Defaults to 6, which corresponds to
a cubic spline as explained in Carty et al. (2017)}

\item{Dmin}{minimum distance (included) to check for significant interactions,
defaults to 0}

\item{Dmax}{maximum distance (included) to check for significant interactions,
defaults to 2e6 or maximum in the data; whichever is minimum.}

\item{ssize}{Distance stratified sampling size. Can decrease for
large chromosomes. Increase recommended if
model fails to converge. Defaults to 0.01.}

\item{ncore}{Number of cores to parallelize. Defaults to 
\code{parallel::detectCores()-1}.}
}
\value{
A valid \code{gi_list} instance with additional \code{mcols(.)} for
each chromosome: pvalue': significance \emph{P}-value, 'qvalue': FDR 
corrected \emph{P}-value, mu': expected counts, 'sdev': modeled standard
deviation of expected counts.
}
\description{
This function finds significant interactions in a HiC-DC readable matrix
and expresses statistical significance of counts through the following with
a parallel implementation (using sockets; compatible with Windows):
'pvalue': significance \emph{P}-value, 'qvalue': FDR corrected 
\emph{P}-value, mu': expected counts, 'sdev': modeled standard deviation
of expected counts.
}
\examples{
gi_list<-generate_binned_gi_list(50e3,chrs='chr22')
gi_list<-add_hic_counts(gi_list,
hic_path=system.file("extdata", "GSE63525_HMEC_combined_example.hic",
package = "HiCDCPlus"))
gi<-HiCDCPlus_parallel(gi_list,ncore=1)
}
