% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopDom.R
\name{TopDom}
\alias{TopDom}
\title{TopDom.R}
\usage{
TopDom(matrix.file, window.size = 5, outFile = NULL, statFilter = TRUE)
}
\arguments{
\item{matrix.file}{string, file address,
Has a structure of N * (N + 3), where N is the number of bins, see Vignette
at https://github.com/HenrikBengtsson/TopDom/blob/0.0.2/docs}

\item{window.size}{number of bins to extend. Defaults to 5}

\item{outFile}{path prefix for TAD annotations to write named across
chromosomes. Defaults to NULL (no file output).}

\item{statFilter}{whether a Wilcoxon rank sum (unpaired) test based filtering
of TAD boundaries would be
done based on a 0.05 \emph{P}-value threshold. Defaults to TRUE.}
}
\value{
A list of TAD annotations per each chromosome
}
\description{
Adapted version of the stable legacy TopDom package 
version 0.0.2 (no longer on CRAN or Bioconductor) 
written by Hanjun Shin(shanjun "at" usc.edu), contributions by 
Harris Lazaris(Ph.D Stduent, NYU), Dr. Gangqing Hu(Staff Scientist, NIH). 
If you're using this function, please cite 
TopDom according to the documentation at 
https://github.com/HenrikBengtsson/TopDom/blob/0.0.2/docs/.
}
